Add basic circular buffer for serial transmissions.
This may be buggy at the moment - it works fine for this basic test loop (sending 0-9 every 20ms), but seems to get stuck if data is queued when the buffer is mostly empty (?).  In that case the transmission of the newly-queued data seems to wait until the next chunk is queued, which is very weird.  Will worry about that later, if needed.