Check UART interrupt enable bits when despatching UART interrupts!
This was the root cause of the transmit buffer problem the other evening; the IOC interrupts from the keyboard scanning logic were also causing a call to the UART Tx ISR if the UART wasn't busy, which would in turn merrily spit out whatever happened to be left over in the buffer from previous transmissions.  Oops.