Use a new, far superior method for injecting keystrokes.
The previous method allowed the injection of one key (or at best, any subset of the keys in a given scan row); the timing was so tight that by the time we checked if a second target row was being scanning, we were too late to successfully inject another key.

This new method runs in constant time using a 256-byte lookup table filled with data to inject; the combined row strobes are used as an index into it to find the first set of column outputs, and then inverted to find the second set.  This allows us to respond to each scan cycle in 14 instructions (plus interrupt latency of 3-5 instructions), and also allow any combination of keys to be chorded since we can respond to all row strobes equally.