Add support for injecting arbitrary key chords.
Now that we have the newer injection routine that doesn't require every key in a chord to be in the same scan row, we can allow any combination of keys to be injected at will.  Although we only need to support two types of chord - Shift-<key> and Code-<key> - and both Shift and Code are in the same scan row, we might as well expose the ability to hold down any two keys.  More than two would probably be excessive and pointless!