Track carriage position in physical units (1/120") instead of characters.
This allows the margins to properly reflect the typewriter's handling of them - they're set as absolute physical positions on the page, and are unaffected by changes to pitch etc.  By keeping track of the carriage position in 1/120" units we can adjust it by half-character cells (for half-backspaces and centered text) without doing gnarly conversions when the pitch or margins change.