Change millisecond timer to start itself aligned to the first active interval.
This avoids the situation where the first tick of a countdown timer is shortened because it was started partway through a timer cycle.  If no other timers are running, the cycle will be realigned to match, resulting in more accurate timing.

Also change setting and reading the countdown value to be atomic with respect to the timer ISR, and have the ISR enable/disable itself based on whether any timers are ccurrently running for neatness's sake.