Handle the LMar/RMar/Mar Rel keys to allow the user to reconfigure margins.
When the users presses one of these keys, the relevant margin is set to the current location (or leftmost location for Mar Rel); this seems to mirror the typewriter's behaviour, which is why we do it!

There may be a subtle different in right-margin handling, though; Mar Rtn seems to supress it for the rest of the line, so TODO investigate this more thoroughly.