Initial implementation of typematic support.
Because the typewriter will happily auto-repeat spaces (or any key after Repeat is held down), we need to emulate the same thing so that we end up with different characters on paper as sent to the host.

Alas getting the timing exactly right has proved tricky but this is a reasonable first approximation.